chore(ButtonWrapper): リファクタリングして型との依存関係を整理する
